Refactor of Tape
Tape completely refactored to be less bloody stupid Johnson in its
architecture.
Cells are now static on screen, and values are rotated through a
separate tape on to them, to save tons of memory.
Also allows large state machines to report instantly, because we don't
need to create interim cells.
Other minor changes included (eg. font-size for labels once we get to 5
characters